rename types/chess-data.ts to types/chess.ts
